utiijnt?

+++ utils/room_data.go
@@ -80,11 +80,11 @@ func GetRoomData(bot structs.BotInterface, userRecordId string) (structs.Room, e
 }
 
 const (
-	TileWidth = int(96 / 2)
+	TileWidth = (96 / 2)
 	// FIX: Use integer arithmetic or explicitly cast the float result.
 	// We'll calculate it using integer arithmetic by multiplying by 587 and dividing by 1000.
 	// 96 * 0.587 = 56.352. (96 * 587) / 1000 = 56.352 (Go truncates the integer division to 56).
-	TileHeight    = int((TileWidth * 587) / 1000 / 2)
+	TileHeight    = (TileWidth * 587) / 1000 / 2)
 	FloorGridSize = 6
 
 	// Estimated center of the room area within the final image canvas.
